@@ -4,38 +4,39 @@
   "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.10.4",
-    "@emotion/styled": "^11.10.4",
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
     "@fortawesome/fontawesome": "^1.1.8",
     "@fortawesome/fontawesome-free-solid": "^5.0.13",
-    "@fortawesome/react-fontawesome": "^0.1.14",
-    "@mui/material": "^5.10.9",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.4.0",
-    "@testing-library/user-event": "^7.2.1",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-router-dom": "^5.1.3",
-    "@types/react-vis": "^1.11.6",
-    "ajv": "^6.12.1",
-    "axios": "^0.25.0",
-    "bootstrap": "^5.1.3",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@mui/material": "^5.12.1",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/react-dom": "^18.0.11",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-vis": "^1.11.11",
+    "ajv": "^8.12.0",
+    "axios": "^1.3.6",
+    "bootstrap": "^5.2.3",
     "check-password-strength": "^2.0.7",
-    "client-zip": "^2.0.0",
+    "client-zip": "^2.3.1",
     "clone-deep": "^4.0.1",
     "file-saver": "^2.0.5",
-    "flat": "^5.0.0",
-    "lodash": "^4.17.19",
-    "natural-orderby": "^2.0.3",
-    "query-string": "^6.9.0",
-    "react": "^17.0.0",
-    "react-bootstrap": "^2.1.2",
-    "react-dom": "^17.0.0",
-    "react-infinite-scroll-component": "^5.0.4",
-    "react-loading-skeleton": "^2.0.1",
-    "react-router-dom": "^6.3.0",
+    "flat": "^5.0.2",
+    "lodash": "^4.17.21",
+    "natural-orderby": "^3.0.2",
+    "papaparse": "^5.4.1",
+    "query-string": "^8.1.0",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.7.4",
+    "react-dom": "^18.2.0",
+    "react-infinite-scroll-component": "^6.1.0",
+    "react-loading-skeleton": "^3.2.1",
+    "react-router-dom": "^6.10.0",
     "react-scripts": "5.0.1",
     "request-promise": "^4.2.6",
-    "sass": "^1.48.0",
+    "sass": "^1.62.0",
     "typescript": "^3.7.4"
   },
   "scripts": {
@@ -53,7 +54,13 @@
     ],
     "rules": {
       "eqeqeq": "off",
-      "no-unused-vars": "off"
+      "no-unused-vars": "off",
+      "array-callback-return": "off",
+      "react-hook/exhaustive-deps": "off",
+      "@typescript-eslint/no-unused-vars": "off",
+      "react/jsx-no-target-blank": "off",
+      "no-redeclare": "off",
+      "react-hooks/exhaustive-deps": "off"
     }
   },
   "browserslist": {
@@ -69,9 +76,11 @@
     ]
   },
   "devDependencies": {
+    "@types/babel__core": "^7.20.0",
     "@types/clone-deep": "^4.0.1",
     "@types/file-saver": "^2.0.5",
-    "@types/react": "^17.0.39",
-    "cypress": "^12.4.0"
+    "@types/papaparse": "^5.3.7",
+    "@types/react": "^18.0.37",
+    "cypress": "^12.10.0"
   }
 }
